1.From Lection 2.Add Calculator 3. div 0 and RequestMapping 4. add Exception 5. Modify lvl_1 6. Add new exception 7. Add tests 8. End.

+++ src/main/java/pro/sky/skypromavenspring/controller/TextController.java
@@ -39,8 +39,8 @@ public class TextController {
     }
 
     @RequestMapping(path = "divide")
-    public String calculatorPageCalcDiv(@RequestParam(value = "num1", required = false) Integer valueOne,
-                                        @RequestParam(value = "num2", required = false) Integer valueTwo) {
+    public String calculatorPageCalcDiv(@RequestParam(value = "num1", required = false) Double valueOne,
+                                        @RequestParam(value = "num2", required = false) Double valueTwo) {
         return textService.calculatorPageCalcDiv(valueOne, valueTwo);
     }
