+++ src/test/java/pro/sky/skypromavenspring/TextServiceParametrizedTests.java
@@ -12,6 +12,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class TextServiceParametrizedTests {
 
     TextService out = new TextService();
+
     public static Stream<Arguments> provideArgumentsPlus() {
         return Stream.of(
                 Arguments.of(4, 7, 11),
@@ -19,6 +20,7 @@ public class TextServiceParametrizedTests {
                 Arguments.of(-4, -7, -11)
         );
     }
+
     public static Stream<Arguments> provideArgumentsMinus() {
         return Stream.of(
                 Arguments.of(4, 7, -3),
@@ -26,6 +28,7 @@ public class TextServiceParametrizedTests {
                 Arguments.of(-4, -7, 3)
         );
     }
+
     public static Stream<Arguments> provideArgumentsMul() {
         return Stream.of(
                 Arguments.of(4, 7, 28),
@@ -33,6 +36,7 @@ public class TextServiceParametrizedTests {
                 Arguments.of(-4, -7, 28)
         );
     }
+
     public static Stream<Arguments> provideArgumentsDiv() {
         return Stream.of(
                 Arguments.of(4, 2, 2.0),
@@ -40,29 +44,31 @@ public class TextServiceParametrizedTests {
                 Arguments.of(-4, -2, 2.0)
         );
     }
+
     @ParameterizedTest
     @MethodSource("provideArgumentsPlus")
-    public void argumentsPlusTest(Integer valueOne, Integer valueTwo, Integer value){
+    public void argumentsPlusTest(Integer valueOne, Integer valueTwo, Integer value) {
         assertEquals("Значение сложения: " + valueOne + " + " + valueTwo + " = " + value,
                 out.calculatorPageCalcPlus(valueOne, valueTwo));
     }
 
     @ParameterizedTest
     @MethodSource("provideArgumentsMinus")
-    public void argumentsMinusTest(Integer valueOne, Integer valueTwo, Integer value){
+    public void argumentsMinusTest(Integer valueOne, Integer valueTwo, Integer value) {
         assertEquals("Значение вычитания: " + valueOne + " - " + valueTwo + " = " + value,
                 out.calculatorPageCalcMinus(valueOne, valueTwo));
     }
+
     @ParameterizedTest
     @MethodSource("provideArgumentsMul")
-    public void argumentsMulTest(Integer valueOne, Integer valueTwo, Integer value){
+    public void argumentsMulTest(Integer valueOne, Integer valueTwo, Integer value) {
         assertEquals("Значение умножения: " + valueOne + " * " + valueTwo + " = " + value,
                 out.calculatorPageCalcMul(valueOne, valueTwo));
     }
 
     @ParameterizedTest
     @MethodSource("provideArgumentsDiv")
-    public void argumentsDivTest(Integer valueOne, Integer valueTwo, double value){
+    public void argumentsDivTest(double valueOne, double valueTwo, double value) {
         assertEquals("Значение деления: " + valueOne + " / " + valueTwo + " = " + value,
                 out.calculatorPageCalcDiv(valueOne, valueTwo));
     }
